Move Proof and Difficulty types to pow module. Difficulty scaling calculation.
