+++ chain/src/pipe.rs
@@ -23,10 +23,10 @@ use chrono::Duration;
 use chain::OrphanBlockPool;
 use core::consensus;
 use core::core::hash::{Hash, Hashed};
-use core::core::target::Difficulty;
 use core::core::verifier_cache::VerifierCache;
 use core::core::{Block, BlockHeader};
 use core::global;
+use core::pow::Difficulty;
 use error::{Error, ErrorKind};
 use grin_store;
 use store;
