+++ chain/tests/test_coinbase_maturity.rs
@@ -27,10 +27,10 @@ use std::sync::{Arc, RwLock};
 
 use chain::types::NoopAdapter;
 use chain::ErrorKind;
-use core::core::target::Difficulty;
 use core::core::transaction;
 use core::core::verifier_cache::LruVerifierCache;
 use core::global::{self, ChainTypes};
+use core::pow::Difficulty;
 use core::{consensus, pow};
 use keychain::{ExtKeychain, Keychain};
 use wallet::libtx::{self, build};
