+++ chain/tests/test_txhashset.rs
@@ -28,8 +28,8 @@ use std::sync::Arc;
 use chain::store::ChainStore;
 use chain::txhashset;
 use chain::types::Tip;
-use core::core::target::Difficulty;
 use core::core::{Block, BlockHeader};
+use core::pow::Difficulty;
 use keychain::{ExtKeychain, Keychain};
 use util::file;
 use wallet::libtx::{build, reward};
