+++ core/src/global.rs
@@ -22,7 +22,7 @@ use consensus::{
 	DIFFICULTY_ADJUST_WINDOW, INITIAL_DIFFICULTY, MEDIAN_TIME_WINDOW, PROOFSIZE,
 	REFERENCE_SIZESHIFT,
 };
-use core::target::Difficulty;
+use pow::Difficulty;
 /// An enum collecting sets of parameters used throughout the
 /// code wherever mining is needed. This should allow for
 /// different sets of parameters for different purposes,
@@ -145,8 +145,11 @@ pub fn coinbase_maturity(height: u64) -> u64 {
 	match *param_ref {
 		ChainTypes::AutomatedTesting => AUTOMATED_TESTING_COINBASE_MATURITY,
 		ChainTypes::UserTesting => USER_TESTING_COINBASE_MATURITY,
-		_ => if height < COINBASE_MATURITY_FORK_HEIGHT { OLD_COINBASE_MATURITY }
-                     else { COINBASE_MATURITY },
+		_ => if height < COINBASE_MATURITY_FORK_HEIGHT {
+			OLD_COINBASE_MATURITY
+		} else {
+			COINBASE_MATURITY
+		},
 	}
 }
