+++ core/src/pow/mod.rs
@@ -37,14 +37,16 @@ extern crate grin_util as util;
 
 pub mod cuckoo;
 mod siphash;
+mod types;
 
 use chrono::prelude::{DateTime, NaiveDateTime, Utc};
 use consensus;
-use core::target::Difficulty;
 use core::{Block, BlockHeader};
 use genesis;
 use global;
 use pow::cuckoo::{Cuckoo, Error};
+
+pub use self::types::*;
 
 /// Validates the proof of work of a given header, and that the proof of work
 /// satisfies the requirements of the header.
@@ -92,8 +94,8 @@ pub fn pow_size(
 		// if we found a cycle (not guaranteed) and the proof hash is higher that the
 		// diff, we're all good
 		if let Ok(proof) = cuckoo::Miner::new(bh, consensus::EASINESS, proof_size, sz).mine() {
-			if proof.to_difficulty() >= diff {
-				bh.pow.proof = proof.clone();
+			bh.pow.proof = proof;
+			if bh.pow.to_difficulty() >= diff {
 				return Ok(());
 			}
 		}
@@ -113,7 +115,6 @@ pub fn pow_size(
 #[cfg(test)]
 mod test {
 	use super::*;
-	use core::target::Difficulty;
 	use genesis;
 	use global;
 
@@ -130,7 +131,7 @@ mod test {
 			global::min_sizeshift(),
 		).unwrap();
 		assert!(b.header.pow.nonce != 310);
-		assert!(b.header.pow.proof.to_difficulty() >= Difficulty::one());
+		assert!(b.header.pow.to_difficulty() >= Difficulty::one());
 		assert!(verify_size(&b.header, global::min_sizeshift()));
 	}
 }
