+++ core/src/ser.rs
@@ -408,8 +408,9 @@ pub struct BinWriter<'a> {
 }
 
 impl<'a> BinWriter<'a> {
+	/// Wraps a standard Write in a new BinWriter
 	pub fn new(write: &'a mut Write) -> BinWriter<'a> {
-		BinWriter{sink: write}
+		BinWriter { sink: write }
 	}
 }
