+++ servers/src/common/adapters.rs
@@ -26,10 +26,10 @@ use std::time::Instant;
 use chain::{self, ChainAdapter, Options, Tip};
 use common::types::{self, ChainValidationMode, ServerConfig, SyncState, SyncStatus};
 use core::core::hash::{Hash, Hashed};
-use core::core::target::Difficulty;
 use core::core::transaction::Transaction;
 use core::core::verifier_cache::VerifierCache;
 use core::core::{BlockHeader, CompactBlock};
+use core::pow::Difficulty;
 use core::{core, global};
 use p2p;
 use pool;
