+++ servers/src/grin/sync.rs
@@ -22,8 +22,8 @@ use std::{cmp, thread};
 use chain;
 use common::types::{Error, SyncState, SyncStatus};
 use core::core::hash::{Hash, Hashed};
-use core::core::target::Difficulty;
 use core::global;
+use core::pow::Difficulty;
 use grin::sync;
 use p2p::{self, Peer, Peers};
 use util::LOGGER;
