+++ servers/src/mining/test_miner.rs
@@ -25,8 +25,8 @@ use chain;
 use common::types::StratumServerConfig;
 use core::core::hash::{Hash, Hashed};
 use core::core::verifier_cache::VerifierCache;
-use core::core::{Block, BlockHeader, Proof};
-use core::pow::cuckoo;
+use core::core::{Block, BlockHeader};
+use core::pow::{cuckoo, Proof};
 use core::{consensus, global};
 use mining::mine_block;
 use pool;
